Pretty-print configurations
Add two new files, config.h and config.c, to handle configuration
structs. Currently, only a configuration-to-string function is
available. In a future commit, the new files will also be responsible
for config comparison (ResTable_config::isBetterThan and friends).

Update the dump command to make use of the new functionality.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>